pyproject.toml
[project]
name = "link_sharing_app"
version = "1.0.0"
description = "Link sharing application backend"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.8.0",
    "waitress>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "coverage>=7.0.0",
    "ruff>=0.8.0",
]

[project.scripts]
dev = "flask:cli"
serve = "waitress:serve"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["link_sharing_app"]

[tool.ruff]
target-version = "py314"
line-length = 88
exclude = [
    ".venv",
    "venv",
    "__pycache__",
    ".git",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "N",
    "UP",
    "B",
    "C4",
    "SIM",
    "RET",
    "ARG",
    "PTH",
    "ERA",
    "PL",
    "PERF",
]
ignore = [
    "E501",
    "PLR0913",
    "PLR2004",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "PLR2004"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
pretty = true
show_error_context = true
show_column_numbers = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "types-flask>=1.1.6",
]
